+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 1.21.0 - 2026-09-17
+
 ### Added
 
 - Add replication destination prefix support, [PR-94](https://github.com/reductstore/reduct-rs/pull/94)